Time 클래스, Double Buffering구현

+++ DuhiEngine_Source/DHDinoObject.h
@@ -20,7 +20,7 @@ namespace dh {
 		float Ypos;
 		float Xspeed;
 		float Yspeed;
-		float Xacc;
+		double Xacc;
 		float Yacc;
 		float height;
